TASK: Extract php 8 specific test classes into separate folder for Neos 7.3 compatibility
